Create SQL Query Analyzer
When we talk about optimizing SQL queries, we are usually looking to improve the performance of the query. This could mean reducing the time it takes to execute the query, reducing the amount of system resources (CPU, memory, disk I/O) required to execute the query, or reducing the amount of data that needs to be read or written during the query. Here,we use the sqlite3 module to connect to a local SQLite database and execute SQL queries against it. The analyze_query function takes a SQL query as input, executes it with EXPLAIN QUERY PLAN and EXPLAIN, and prints the query plan and execution details. The main function prompts the user for a SQL query and calls the analyze_query function with the user's query. When you run the program, you can enter a SQL query and the program will analyze it for potential optimizations.